move pydantic models to separate config file
